requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
pyTelegramBotAPI==4.21.0
python-dotenv==1.0.1
requests==2.32.3
telegram==0.0.1
urllib3==2.2.2
yarl==1.9.4
